Extract assist-chat out of voice-command-dialog
